Expand incremental compilation unit test coverage to include Roslyn tracking.
