+++ src/libraries/System.Text.Json/tests/System.Text.Json.SourceGeneration.Unit.Tests/JsonSourceGeneratorTests.cs
@@ -663,172 +663,5 @@ namespace System.Text.Json.SourceGeneration.UnitTests
             CompilationHelper.CheckDiagnosticMessages(DiagnosticSeverity.Warning, generatorDiags, Array.Empty<(Location, string)>());
             CompilationHelper.CheckDiagnosticMessages(DiagnosticSeverity.Error, generatorDiags, Array.Empty<(Location, string)>());
         }
-
-        [Theory]
-        [MemberData(nameof(GetCompilationHelperFactories))]
-        public static void CompilingTheSameSourceResultsInEqualModels(Func<Compilation> factory)
-        {
-            JsonSourceGeneratorResult result1 = CompilationHelper.RunJsonSourceGenerator(factory());
-            JsonSourceGeneratorResult result2 = CompilationHelper.RunJsonSourceGenerator(factory());
-
-            if (result1.SourceGenModel is null)
-            {
-                Assert.Null(result2.SourceGenModel);
-            }
-            else
-            {
-                Assert.NotSame(result1.SourceGenModel, result2.SourceGenModel);
-
-                Assert.Equal(result1.SourceGenModel, result2.SourceGenModel);
-                Assert.Equal(result1.SourceGenModel.GetHashCode(), result2.SourceGenModel.GetHashCode());
-            }
-        }
-
-        [Fact]
-        public static void CompilingEquivalentSourcesResultsInEqualModels()
-        {
-            string source1 = """
-                using System;
-                using System.Text.Json.Serialization;
-                
-                namespace Test
-                {
-                    [JsonSerializable(typeof(MyPoco))]
-                    public partial class JsonContext : JsonSerializerContext { }
-                
-                    public class MyPoco
-                    {
-                        public string MyProperty { get; set; } = 42;
-                    }
-                }
-                """;
-
-            string source2 = """
-                using System;
-                using System.Text.Json.Serialization;
-                
-                namespace Test
-                {
-                    // Same as above but with different implementation
-                    public class MyPoco
-                    {
-                        public string MyProperty
-                        {
-                            get => -1;
-                            set => throw new NotSupportedException();
-                        }
-                    }
-
-                    // Changing location should produce identical SG model when no diagnostics are emitted.
-                    [JsonSerializable(typeof(MyPoco))]
-                    public partial class JsonContext : JsonSerializerContext { }
-                }
-                """;
-
-            JsonSourceGeneratorResult result1 = CompilationHelper.RunJsonSourceGenerator(CompilationHelper.CreateCompilation(source1));
-            JsonSourceGeneratorResult result2 = CompilationHelper.RunJsonSourceGenerator(CompilationHelper.CreateCompilation(source2));
-            Assert.Empty(result1.Diagnostics);
-            Assert.Empty(result2.Diagnostics);
-
-            Assert.NotSame(result1.SourceGenModel, result2.SourceGenModel);
-            Assert.Equal(result1.SourceGenModel, result2.SourceGenModel);
-            Assert.Equal(result1.SourceGenModel.GetHashCode(), result2.SourceGenModel.GetHashCode());
-        }
-
-        [Fact]
-        public static void CompilingDifferentSourcesResultsInUnequalModels()
-        {
-            string source1 = """
-                using System;
-                using System.Text.Json.Serialization;
-                
-                namespace Test
-                {
-                    [JsonSerializable(typeof(MyPoco))]
-                    public partial class JsonContext : JsonSerializerContext { }
-                
-                    public class MyPoco
-                    {
-                        public string MyProperty { get; set; } = 42;
-                    }
-                }
-                """;
-
-            string source2 = """
-                using System;
-                using System.Text.Json.Serialization;
-                
-                namespace Test
-                {
-                    [JsonSerializable(typeof(MyPoco))]
-                    public partial class JsonContext : JsonSerializerContext { }
-                
-                    public class MyPoco
-                    {
-                        public string MyProperty { get; } = 42; // same, but missing a getter
-                    }
-                }
-                """;
-
-            JsonSourceGeneratorResult result1 = CompilationHelper.RunJsonSourceGenerator(CompilationHelper.CreateCompilation(source1));
-            JsonSourceGeneratorResult result2 = CompilationHelper.RunJsonSourceGenerator(CompilationHelper.CreateCompilation(source2));
-            Assert.Empty(result1.Diagnostics);
-            Assert.Empty(result2.Diagnostics);
-
-            Assert.NotEqual(result1.SourceGenModel, result2.SourceGenModel);
-        }
-
-        [Theory]
-        [MemberData(nameof(GetCompilationHelperFactories))]
-        public static void SourceGenModelDoesNotEncapsulateSymbolsOrCompilationData(Func<Compilation> factory)
-        {
-            JsonSourceGeneratorResult result = CompilationHelper.RunJsonSourceGenerator(factory());
-            WalkObjectGraph(result.SourceGenModel);
-
-            static void WalkObjectGraph(object obj)
-            {
-                var visited = new HashSet<object>();
-                Visit(obj);
-
-                void Visit(object? node)
-                {
-                    if (node is null || !visited.Add(node))
-                    {
-                        return;
-                    }
-
-                    Assert.False(node is Compilation or ISymbol);
-
-                    Type type = node.GetType();
-                    if (type.IsPrimitive || type.IsEnum || type == typeof(string))
-                    {
-                        return;
-                    }
-
-                    if (node is IEnumerable collection and not string)
-                    {
-                        foreach (object? element in collection)
-                        {
-                            Visit(element);
-                        }
-
-                        return;
-                    }
-
-                    foreach (FieldInfo field in type.GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance))
-                    {
-                        object? fieldValue = field.GetValue(node);
-                        Visit(fieldValue);
-                    }
-                }
-            }
-        }
-
-        public static IEnumerable<object[]> GetCompilationHelperFactories()
-        {
-            return typeof(CompilationHelper).GetMethods(BindingFlags.Static | BindingFlags.Public)
-                .Where(m => m.ReturnType == typeof(Compilation) && m.GetParameters().Length == 0)
-                .Select(m => new object[] { Delegate.CreateDelegate(typeof(Func<Compilation>), m) });
-        }
     }
 }
