+++ src/libraries/System.Text.Json/tests/System.Text.Json.SourceGeneration.Unit.Tests/System.Text.Json.SourceGeneration.Unit.Tests.targets
@@ -14,6 +14,7 @@
   <ItemGroup>
     <Compile Include="CompilationHelper.cs" />
     <Compile Include="JsonSourceGeneratorDiagnosticsTests.cs" />
+    <Compile Include="JsonSourceGeneratorIncrementalTests.cs" />
     <Compile Include="JsonSourceGeneratorTests.cs" />
     <Compile Include="TypeWrapperTests.cs" />
   </ItemGroup>
